Refactor project selector.
